Add refactored API calls logic to contractor pages and compomnents
