+++ client/src/components/index.js
@@ -1,4 +1,3 @@
 import Navbar from './Navbar';
-import InfoPopup from './InfoPopup';
 
-export { Navbar, InfoPopup };
+export { Navbar };
