+++ client/src/pages/contractor/ContractorProfile.js
@@ -3,7 +3,7 @@ import { Layout, Typography, Input, Button, Card, Form } from 'antd';
 import {
   fetchContractorProfile,
   saveContractorProfile,
-} from '../../api/firebaseContractor';
+} from '../../api/contractor/profile';
 
 const { Content } = Layout;
 const { Title } = Typography;
@@ -28,15 +28,24 @@ const ContractorProfile = () => {
 
   useEffect(() => {
     const loadProfile = async () => {
-      await fetchContractorProfile(uid, form, defaultFields);
-      setLoading(false);
+      try {
+        await fetchContractorProfile(uid, form, defaultFields);
+      } catch (err) {
+        console.error('Failed to fetch contractor profile', err);
+      } finally {
+        setLoading(false);
+      }
     };
     loadProfile();
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
   const handleSubmit = async (values) => {
-    await saveContractorProfile(uid, values, email, defaultFields);
+    try {
+      await saveContractorProfile(uid, values, email, defaultFields);
+    } catch (err) {
+      console.error('Failed to save contractor profile', err);
+    }
   };
 
   return (
